Move the input state outside
