+++ jvm/src/jmh/kotlin/com/github/jcornaz/collekt/IterationBenchmark.kt
@@ -1,0 +1,13 @@
+package com.github.jcornaz.collekt
+
+import org.openjdk.jmh.annotations.Benchmark
+import org.openjdk.jmh.infra.Blackhole
+
+
+open class IterationBenchmark {
+
+    @Benchmark
+    fun measureIteration(input: PersistentInput, blackhole: Blackhole) {
+        input.list.forEach { blackhole.consume(it) }
+    }
+}
